Added some project boilerplate
Including but not limited to:
- A Makefile do speed up testing and development environment setup
- A skeleton for unit tests

I took the opinionated decision to use the nose test runner (because
it's the best in my humble opinion), but of course any other test runner
can be used instead since the tests only use the standard
unittest module.

This commit is provided free of charge and without any warranty, and the
code it encompasses is hereby released to the public domain (use as you see
fit for any purpose).